+++ src/test/java/de/akquinet/jbosscc/guttenbase/hints/TableOrderHintTest.java
@@ -1,11 +1,6 @@
 package de.akquinet.jbosscc.guttenbase.hints;
 
-import java.util.Comparator;
-
 import org.junit.Before;
-
-import de.akquinet.jbosscc.guttenbase.mapping.TableOrderComparatorFactory;
-import de.akquinet.jbosscc.guttenbase.meta.TableMetaData;
 
 /**
  * Test a schema migration where table ordering is customized...
@@ -23,21 +18,6 @@ public class TableOrderHintTest extends AbstractHintTest {
 
 	@Before
 	public void setup() throws Exception {
-		_connectorRepository.addConnectorHint(TARGET, new TableOrderHint() {
-			@Override
-			public TableOrderComparatorFactory getValue() {
-				return new TableOrderComparatorFactory() {
-					@Override
-					public Comparator<TableMetaData> createComparator() {
-						return new Comparator<TableMetaData>() {
-							@Override
-							public int compare(final TableMetaData o1, final TableMetaData o2) {
-								return o1.hashCode() - o2.hashCode(); // Just for demonstration purposes...
-							}
-						};
-					}
-				};
-			}
-		});
+		_connectorRepository.addConnectorHint(TARGET, new RandomTableOrderHint());
 	}
 }