+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+What's new in Version 1.3
+============================
+- Handling case in table names, etc. improved
+- TableOrderTool computes DROP/CREATE order for tables
+
 What's new in Version 1.2
 ============================
 - Many bug fixes and enhancements